project: allow use of stable rust
This project currently requires nightly for two reasons:

1. The use of unstable feature `rustc_private` in
   `longfi-sys/build.rs`
2. Use of unstable feature `lang_items` in
   `examples/stm32l0x2/main.rs`

Neither one is required.

(1) _was_ necessary because `longfi-sys/Cargo.toml` did not delcare
`cc` as a depdendency. As a result, the compiler assumed that we
wanted `rustc`'s own `cc` crate. Direct use of Rust's internal
dependencies is an unstalbe feature and does require `rustc_private`.

(2) was not needed at all, and is likeley just a result of starting
with an outdated RTFM example, as RTFM one did require `lang_items`
but removed the need sometime in 2018.